+++ HISTORY.rst
@@ -10,7 +10,11 @@ History
 
 * Dropped support for Django<1.8, django CMS<3.2.
 * Added liveblog application.
+* Refactored plugin filters: by default only data for current site are now shown.
+* Added global and per site posts count to BlogCategory.
+* Added option to hide empty categories from menu.
 * Added standalone documentation at https://djangocms-blog.readthedocs.io.
+* Enabled cached version of BlogLatestEntriesPlugin.
 
 ******************
 0.8.5 (2016-06-26)